# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DumbPad API",
    "version": "1.0.8",
    "description": "DumbPad content-management API. Legacy deployments authenticate with a PIN; AUTH_V2_ENABLED deployments use an opaque browser session or a scoped API token. API tokens cannot access authentication or data-management routes."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "cookieAuth": [] }, { "sessionCookieAuth": [] }, { "apiTokenBearer": [] }, { "pinBearer": [] }],
  "paths": {
...
    "/api/auth/setup/start": {
      "post": { "security": [], "summary": "Begin one-time V2 password and TOTP setup", "responses": { "200": { "description": "Temporary setup id and TOTP secret" }, "401": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/api/auth/setup/confirm": {
      "post": { "security": [], "summary": "Confirm V2 TOTP setup and return one-time recovery codes", "responses": { "200": { "description": "Initialized session and recovery codes" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" } } }
    },
    "/api/auth/login": {
      "post": { "security": [], "summary": "Create an opaque V2 browser session", "responses": { "200": { "description": "Session cookie and expiry metadata" }, "401": { "$ref": "#/components/responses/Error" }, "429": { "$ref": "#/components/responses/Error" } } }
    },
    "/api/auth/recovery/start": {
      "post": { "security": [], "summary": "Validate a recovery code and begin TOTP replacement", "responses": { "200": { "description": "Temporary recovery id and new TOTP secret" }, "401": { "$ref": "#/components/responses/Error" } } }
    },
    "/api/auth/recovery/confirm": {
      "post": { "security": [], "summary": "Confirm a replacement TOTP secret using a one-time recovery code", "responses": { "200": { "description": "Recovered session" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" } } }
    },
    "/api/auth/elevate": {
      "post": { "security": [{ "sessionCookieAuth": [] }], "summary": "Verify TOTP for a short high-risk-operation window", "responses": { "200": { "description": "Elevation expiry" }, "401": { "$ref": "#/components/responses/Error" } } }
    },
    "/api/auth/logout": {
      "post": { "security": [{ "sessionCookieAuth": [] }], "summary": "Revoke the current V2 browser session", "responses": { "200": { "description": "Session revoked" } } }
    },
    "/api/auth/api-tokens": {
      "get": { "security": [{ "sessionCookieAuth": [] }], "summary": "List scoped API token metadata after elevation", "responses": { "200": { "description": "Token metadata" }, "403": { "$ref": "#/components/responses/Error" } } },
      "post": { "security": [{ "sessionCookieAuth": [] }], "summary": "Create a scoped API token after elevation", "responses": { "201": { "description": "One-time plaintext token and metadata" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/api/auth/api-tokens/{tokenId}": {
      "delete": { "security": [{ "sessionCookieAuth": [] }], "summary": "Revoke a scoped API token after elevation", "responses": { "200": { "description": "Token revoked" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } }
    },
    "/api/auth/devices": {
      "get": { "security": [{ "sessionCookieAuth": [] }], "summary": "List trusted-device metadata after elevation", "responses": { "200": { "description": "Trusted devices" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/api/auth/devices/{deviceId}": {
      "delete": { "security": [{ "sessionCookieAuth": [] }], "summary": "Revoke a trusted device after elevation", "responses": { "200": { "description": "Device and linked sessions revoked" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "dumbpad_auth" },
      "sessionCookieAuth": { "type": "apiKey", "in": "cookie", "name": "dumbpad_auth_session" },
      "apiTokenBearer": { "type": "http", "scheme": "bearer", "description": "Scoped V2 content API token; cannot access auth or data-management routes." },
      "pinBearer": { "type": "http", "scheme": "bearer", "description": "Legacy DumbPad PIN for trusted automation only. Disabled in V2." }
    },
...
    "responses": { "Error": { "description": "Request failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } },
    "schemas": {
      "Error": { "type": "object", "properties": { "error": { "type": "string" }, "currentVersion": { "type": "integer" } }, "required": ["error"] },
...
    }
  }
}